Clean and standardize data

Cleans up messy data frames by performing several operations. These Include cleaning of column names, detecting and removing duplicates, empty records and columns, constant columns, replacing missing values by NA, converting character columns into dates when they contain a certain number of date values, and detecting subject IDs with wrong formats.

Usage

clean_data(
  data,
  params = list(standardize_column_names = list(keep = NULL, rename = NULL),
    replace_missing_values = list(target_columns = NULL, na_strings =
    cleanepi::common_na_strings), remove_duplicates = list(target_columns = NULL,
    rm_empty_rows = TRUE, rm_empty_cols = TRUE, rm_constant_cols = TRUE),
    standardize_dates = list(target_columns = NULL, error_tolerance = 0.4, format = NULL,
    timeframe = NULL, orders = list(world_named_months = c("Ybd", "dby"),
    world_digit_months = c("dmy", "Ymd"), US_formats = c("Omdy", "YOmd")), 
    
    modern_excel = TRUE), standardize_subject_ids = list(target_columns = "id", prefix =
    NULL, suffix = NULL, range = NULL, nchar = NULL), dictionary = NULL, to_numeric =
    NULL)
)

Arguments

data

The input data frame or linelist

params

A list of parameters that define what cleaning operations will be applied on the input data. Possible values are:

keep: A vector of column names to be kept as they appear in the original data. All column names will be standardized if this is NULL (the default value).

replace_missing_values: A list of parameters to be used when replacing the missing values by NA. These parameters are the inputs for the replace_missing_values() function.

remove_duplicates: A list of arguments that defines the list of columns to be considered when looking for duplicates. It also contains arguments that determine how constant rows and columns will be handled. They are the input values for the remove_duplicates() function.

standardize_date: A list of parameters that will be used to standardize the date values from the input data. They represent the input values for the standardize_dates() function.

standardize_subject_ids: A list of parameters that are needed to check the IDs that comply with the expect format. These arguments are the input values of the check_subject_ids().

to_numeric: a vector of column names. When provided, the values in these columns will be converted into numeric.

dictionary: A data frame that will be used to substitute the current values in the specified columns the those in the dictionary. It is the main argument for the clean_using_dictionary() function.

Value

The cleaned input date according to the user-specified parameters. This is associated with a data cleaning report that can be accessed using attr(cleaned_data, "report")

Examples

# Parameters for column names standardization
standardize_col_names <- list(keep = NULL, rename = NULL)

# parameters to remove constant columns, empty rows and columns
remove_cte <- list(cutoff = 1)

# Parameters for substituting missing values with NA:
replace_missing_values <- list(target_columns = NULL, na_strings = "-99")

# Parameters for duplicates removal across all columns
# (target_columns = NULL)
remove_duplicates <- list(target_columns   = NULL,
                          rm_empty_rows    = TRUE,
                          rm_empty_cols    = TRUE,
                          rm_constant_cols = TRUE)

# Parameters for dates standardization
standardize_date <- list(target_columns  = NULL,
                         error_tolerance = 0.4,
                         format          = NULL,
                         timeframe       = as.Date(c("1973-05-29",
                                                     "2023-05-29")),
                         orders          = list(
                           world_named_months = c("Ybd", "dby"),
                           world_digit_months = c("dmy", "Ymd"),
                           US_formats         = c("Omdy", "YOmd")
                         ),
                         modern_excel    = TRUE)

# Parameters for subject IDs standardization
standardize_subject_ids <- list(target_columns = "study_id",
                                prefix         = "PS",
                                suffix         = "P2",
                                range          = c(1, 100),
                                nchar          = 7)

# to_numeric = "sex" the 'sex' column will be converted into numeric

# dictionary = NULL the dictionary-based cleaning will not be performed here

cleaned_data <- clean_data(
  data   = readRDS(system.file("extdata", "test_df.RDS",
                               package = "cleanepi")),
  params = list(
    standardize_column_names = standardize_col_names,
    remove_constant          = remove_cte,
    replace_missing_values   = replace_missing_values,
    remove_duplicates        = remove_duplicates,
    standardize_date         = standardize_date,
    standardize_subject_ids  = standardize_subject_ids,
    to_numeric               = "sex",
    dictionary               = NULL
  )
)
#> 
#> cleaning column names
#> replacing missing values with NA
#> removing the constant columns, empty rows and columns
#> removing duplicated rows
#> standardising date columns
#> checking subject IDs format
#> Warning: Detected incorrect subject ids at lines: 5, 7, 3
#> Use the correct_subject_ids() function to adjust them.
#> 
#> converting sex into numeric