cleanepi: Clean and Standardize Epidemiological Data

Cleaning and standardizing tabular data package, tailored specifically for curating epidemiological data. It streamlines various data cleaning tasks that are typically expected when working with datasets in epidemiology. It returns the processed data in the same format, and generates a comprehensive report detailing the outcomes of each cleaning task.

See also

Author

Maintainer: Karim Mané karim.mane@lshtm.ac.uk (ORCID)

Authors:

Abdoelnaser Degoot abdoelnaser-mahmood.degoot@lshtm.ac.uk (ORCID)

Bankolé Ahadzie Bankole.Ahadzie@lshtm.ac.uk

Nuredin Mohammed Nuredin.Mohammed@lshtm.ac.uk

Bubacarr Bah Bubacarr.Bah1@lshtm.ac.uk (ORCID)

Other contributors:

Thibaut Jombart thibautjombart@gmail.com (Thibault contributed in development of date_guess().) [contributor]

Hugo Gruson hugo@data.org (ORCID) [contributor, reviewer]

Pratik R. Gupte pratik.gupte@lshtm.ac.uk (ORCID) [reviewer]

James M. Azam james.azam@lshtm.ac.uk (ORCID) [reviewer]

Joshua W. Lambert joshua.lambert@lshtm.ac.uk (ORCID) [reviewer]

Chris Hartgerink chris@data.org (ORCID) [reviewer]

Andree Valle-Campos avallecam@gmail.com [reviewer, contributor]

London School of Hygiene and Tropical Medicine, LSHTM (00a0jsq62) [copyright holder]

data.org [funder]