Try and guess dates from a characters

Note that THIS FEATURE IS STILL EXPERIMENTAL: we strongly recommend checking a few converted dates manually. This function tries to extract dates from a character vector or a factor. It treats each entry independently, using regular expressions to detect if a date is present, its format, and if successful it converts that entry to a standard Date with the Ymd format (e.g. 2018-01-21). Entries which cannot be processed result in NA. An error threshold can be used to define the maximum number of resulting NA (i.e. entries without an identified date) that can be tolerated. If this threshold is exceeded, the original vector is returned.

Usage

date_guess(
  x,
  quiet = TRUE,
  modern_excel = TRUE,
  orders = list(world_named_months = c("Ybd", "dby"), world_digit_months = c("dmy",
    "Ymd"), US_formats = c("Omdy", "YOmd"))
)

Arguments

x

A character vector or a factor

quiet

A logical indicating if messages should be displayed to the console (TRUE, default); set to FALSE to silence messages

modern_excel

When parsing dates from excel, some dates are stored as integers. Modern versions of Excel represent dates as the number of days since 1900-01-01, but pre-2011 Excel for OSX have the origin set at 1904-01-01. If this parameter is TRUE (default), then this assumes that all numeric values represent dates from either a Windows version of Excel or a 2011 or later version of Excel for OSX. Set this parameter to FALSE if the data came from an OSX version of Excel before 2011.

orders

The date codes for fine-grained parsing of dates. This allows for parsing of mixed dates. If a list is supplied, that list will be used for successive tries in parsing. Default orders are:

list(
  world_named_months = c("Ybd", "dby"),
  world_digit_months = c("dmy", "Ymd"),
  US_formats         = c("Omdy", "YOmd")
)

Value

A list of following two elements: a vector of the newly reformatted dates and a data frame with the date values that were converted from more than one format. If all values comply with only one format, the later element will be NULL.

Author

Thibaut Jombart, Zhian N. Kamvar

Examples

if (FALSE) {
# Mixed format date -----------------------------------------

guess_dates(c("03 Jan 2018", "07/03/1982", "08/20/85")) # default

# Prioritizing specific date formats ------------------------
#
# The default orders prioritize world date ordering over American-style.
orders <- list(
  world_named_months = c("Ybd", "dby"),
  world_digit_months = c("dmy", "Ymd"),
  US_formats         = c("Omdy", "YOmd")
)

# if you want to prioritize American-style dates with numeric months, you
# can switch the second and third elements of the default orders

print(us_ord <- orders[c(1, 3, 2)])
date_guess(c("03 Jan 2018", "07/03/1982", "08/20/85"), orders = us_ord)

# Handling dates with time formats --------------------------
#
# If you have a format with hours, minutes and seconds, you can also add that
# to the list of formats. Note, however, that this function will drop levels
# below day.

print(orders$ymdhms <- c("Ymdhms", "Ymdhm"))

date_guess(c("2014_04_05_23:15:43", "03 Jan 2018",
"07/03/1982", "08/20/85"), orders = orders)

# Handling  missing and nonsense data -----------------------
#
# date_guess can handle messy dates and tolerate missing data

x <- c("01-12-2001", "male", "female", "2018-10-18", NA, NA, "2018_10_17",
       "43391", "2018 10 19", "// 24/12/1989", "this is 24/12/1989!",
       "RECON NGO: 19 Sep 2018 :)", "6/9/11", "10/10/10")

date_guess(x)
}