Clean data

this function is used to clean up messy data frames. This include the cleaning of column names, detecting and removing of duplicates, empty records and columns, constant columns, replacing missing values by NA, converting character columns into Date when they contain a certain number of Date values, detecting subject IDs with wrong format

Usage

clean_data(
  data,
  params = list(remove_duplicates = FALSE, target_columns = NULL, replace_missing = TRUE,
    na_comes_as = NULL, check_timeframe = FALSE, timeframe = NULL, error_tolerance = 0.5,
    subject_id_col_name = NULL, subject_id_format = NULL, prefix = "PS", suffix = "P2",
    range = c(1L, 100L))
)

Arguments

data

the input data frame

params

a list of parameters that define what cleaning operations will be applied on the input data. Possible parameters are:

remove_duplicates: whether to remove duplicated records or not. If TRUE, the remove argument of the remove_duplicate() function will automatically be set to -1 i.e. to keep only the first instance of duplicated rows. When the user only needs to detect duplicated rows in the dataset, use the find_duplicates() function.

target_columns: a vector of columns names or indices to consider when looking for duplicates. When the input data is a linelist object, this parameter can be set to tags if you wish to look for duplicates across the tagged variables only. Only used when remove_duplicates=TRUE

replace_missing: whether to replace the missing value characters with NA or not. default is FALSE

na_comes_as: the characters that represent the missing values in the data frame. Only used when replace_missing=TRUE

check_timeframe: a logical to determine whether to check if the dates fall under the given time frame of not. default: FALSE

timeframe: a vector of 2 elements of Date class that specifies the first and last date. If provided, all Dates in the data frame must be within this range or set to NA during the cleaning.

error_tolerance: a number between 0 and 1 indicating the proportion of entries which cannot be identified as dates to be tolerated; if this proportion is exceeded, the original vector is returned, and a message is issued; defaults to 0.1 (10 percent)

subject_id_col_name: the name of the column in the data frame with the subject IDs

subject_id_format: the expected subject format

prefix: the prefix used in the subject IDs

suffix: the prefix used in the subject IDs

range: a vector with the range of numbers in the sample IDs

Value

a list of the following 2 elements:

data: the cleaned data frame according to the user-specified parameters

report: a list with the information about the effects of the the cleaning steps

Details

If check_timeframe = TRUE and timeframe = NULL, the timeframe will be today's date and the same date 50 years before.

in clean_data(), duplicated rows will be identified across the user-specified or all columns. Once detected, all occurrences of the duplicated rows will be removed except the first. If you only need to find and remove specific duplicates, use the find_duplicates() then remove_duplicates() functions.

Examples

cleaned_data <- clean_data(
  data   = readRDS(system.file("extdata", "test_df.RDS",
                               package = "cleanepi")),
  params = list(
    remove_duplicates   = TRUE,
    target_columns      = NULL,
    replace_missing     = TRUE,
    na_comes_as         = "-99",
    check_timeframe     = TRUE,
    timeframe           = as.Date(c("1973-05-29", "2023-05-29")),
    error_tolerance     = 0.5,
    subject_id_col_name = "study_id",
    subject_id_format   = "PS000P2",
    prefix              = "PS",
    suffix              = "P2",
    range               = c(1, 100)))
#> 
#> cleaning column names
#> replacing missing values with NA
#> removing empty rows and columns
#> removing constant columns
#> standardising date columns
#> checking for subject IDs uniqueness
#> removing duplicated rows
#> checking subject IDs format