Unnest an element of the data cleaning report

Unnest an element of the data cleaning report

Usage

unnest_report(report, what, ...)

Arguments

report

An object of type <list>

what

A <character> with the name of the specific data cleaning report which would be displayed. The possible values are:

incorrect_date_sequence

To display rows with the incorrect date sequences

colnames

To display the column names before and after cleaning

converted_into_numeric

To display the names of the columns that that have been converted into numeric

date_standardization

To display rows in the cleaned data with date values that are outside of the specified time frame, and rows with date values that comply with multiple formats

misspelled_values

To display the detected misspelled values

removed_duplicates

To display the duplicated rows that have been removed

found_duplicates

To display the duplicated rows

constant_data

To display the constant data i.e. constant columns, empty rows and columns

missing_values_replaced_at

To display the names of the columns where the missing value strings have been replaced with NA

incorrect_subject_id

To display the missing, duplicated and invalid subject subject IDs

scanning_result

To display the output of the scan_data() function

...

Any other extra argument

Value

The input object where the specified element has been unnested and removed.